requirements.txt
accelerate==1.12.0
psutil==7.2.2
aiosmtplib==5.1.0
beautifulsoup4==4.14.3
datasets==3.3.2
fastapi==0.129.2
huggingface_hub==0.29.1
JapaneseTokenizer==1.6
JapaneseTokenizer==1.6
jreadability==1.1.3
kani==1.8.0
konoha==5.6.0
matplotlib==3.10.8
numpy==2.4.2
openai==2.21.0
pandas==3.0.1
python-dotenv==1.2.1
python_Levenshtein==0.27.3
python-multipart==0.0.22
regex==2024.11.6
Requests==2.32.5
scipy==1.17.0
soundfile==0.13.1
stanza==1.10.1
tiktoken==0.9.0
torch==2.6.0
transformers==4.49.0
utils==1.0.2
uvicorn==0.41.0
